requirements.txt
django==latest
djangorestframework
markdown
django-filter
mysqlclient
redis
SpeechRecognition
pocketsphinx
fuzzywuzzy
python-Levenshtein